pyproject.toml
[project]
name = "t-fastapi"
version = "0.1.0"
description = "FastAPI project template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich==0.7.2",
    "apscheduler==3.10.4",
    "fastapi==0.111.0",
    "loguru==0.7.2",
    "passlib[bcrypt]==1.7.4",
    "pydantic-settings==2.2.1",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "redis==5.0.4",
    "tortoise-orm[aiomysql]==0.21.3",
    "uvicorn[standard]==0.29.0",
]

[dependency-groups]
dev = [
    "httpx==0.27.0",
    "mypy>=1.20.2",
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "ruff>=0.15.12",
]

# ── aerich 数据库迁移 ────────────────────────────────────────
[tool.aerich]
tortoise_orm = "core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

# ── pytest ───────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# ── Ruff：lint + 格式化 + import 排序 ───────────────────────
[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
    "N",
]
ignore = [
    "E501",   # 行长度由 line-length 统一控制
    "B008",   # FastAPI Depends() 必须写在参数默认值里，属于框架约定
    "N802",   # 配置类的 @property 用大写命名是有意为之
    "UP046",  # Pydantic 对 PEP 695 新泛型语法支持尚不稳定
]

[tool.ruff.lint.isort]
# 告诉 isort 哪些是自己项目的模块（排序时会归为 first-party 组）
known-first-party = ["core", "api", "models", "schemas", "services", "tasks", "utils"]

# ── Mypy：类型检查 ───────────────────────────────────────────
[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true   # 第三方库缺少 stub 时不报错
disallow_untyped_defs = true    # 函数必须有类型注解
warn_return_any = true
warn_unused_ignores = true